I think I got this bitch workin'
So, this is long but necessary since the code is a little confusing. I'm
praying this commit doesn't fuck up your build. Right now what happens,
is in MainActivity, we create a WifiManager and create a new
BroadcastManager with it as the argument. Then, we send a broadcast
request, and wait for a response. after 1000ms we timeout, on the
assumption we will have received all responses at that point. Right now
the udp packet contains nothing of real interest, but I'll put that in
soon. If you encounter an error that basically won't even run the app,
make sure you do Run As > Run Configurations and make sure you're
running MapShare under android application, as BroadcastManager
registers as a java application.