Signal Strength / Bit Error Rate
So, it's in the comments in the code, but to summarize: I added a
function peer_request that when run from a phone will use context (that
is passed into BroadcastManager constructor) to listen for the signal
strength. I have no idea if the context I pass it will be correct or if
it will work, but it's a start.